services/transcriber-service/requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
av==14.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
ctranslate2==4.6.0
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
faster-whisper==1.1.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
grpcio==1.71.0
grpcio-reflection==1.71.0
grpcio-tools==1.71.0
hf-xet==1.1.0
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
mutagen==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.5
onnxruntime==1.21.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.22.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tinycss2==1.4.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
yt-dlp==2025.5.6.232932.dev0
